requirements.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0

# Spiking Neural Networks
snntorch>=0.6.0

# Orbital mechanics and astronomy
poliastro>=0.17.0
astropy>=5.0
sgp4>=2.0
pyephem>=4.1

# Atmospheric models (optional)
nrlmsise00>=0.1.0

# Data handling
pandas>=1.3.0
h5py>=3.7.0
zarr>=2.12.0

# Experiment tracking and logging
mlflow>=2.0.0
wandb>=0.15.0
tensorboard>=2.10.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0
tqdm>=4.64.0
click>=8.0.0

# Testing and development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Jupyter notebooks (for experiments)
jupyter>=1.0.0
ipykernel>=6.15.0
ipywidgets>=8.0.0

# Docker and deployment (optional)
docker>=6.0.0